Feat : `SnsRecyclerViewAdapter.kt`, `postDTO.kt`
- Raiting bar를 Sns에 띄움에 따라 xml 파일 및 kotlin 파일 수정
- 별점을 sns에 띄울 수 있도록 기능 추가

+++ app/src/main/java/com/a2022sunrinhackathon/activity/AddPostActivity.kt
@@ -80,7 +80,7 @@ class AddPostActivity : AppCompatActivity() {
                 postDTO.imageUrl = uri.toString()
                 postDTO.comment = binding.commentEditText.text.toString()
                 postDTO.address = binding.addressEditText.text.toString()
-                postDTO.rating = binding.addRating.rating.toLong()
+                postDTO.rating = binding.addRating.rating.toFloat()
                 postDTO.timeStamp = System.currentTimeMillis()
 
                 db.collection("posts")
